feat: 동아리 개설 신청서 데이터 전달
Related to #30

+++ src/main/java/environment/project/service/ClubApplyService.java
@@ -10,7 +10,7 @@ public interface ClubApplyService {
 
     int updateApproveYn(ClubApplyDTO clubApplyDTO);
 
-    Long insertClubApply(ClubApplyDTO clubApplyDTO);
+    int insertClubApply(ClubApplyDTO clubApplyDTO);
 
     ClubApplyDTO getApplicationByApplyNo(Long clubApplyNo);
 }
